Re-Refactor Facade.
Re-refactored facade after problems arose. Rewrote some nominal tests. Errors being thrown will be recorded and discussed with team.

Also during this build I built the library such that we now have access to the working exploratory methods in exp.i. These can be subbed in for the facade methods that are currently in a different form than the rest of the repo.